Pulling in recent updates (#20)
* Started breaking out this to be a bit more modular.

Here I've craeted `calculation_methods` attribute. This is a dictionary of functions
which is populated depending upon the specified plot toggle values.  This will make
it much easier for the user to add new properties to calculate.

* Split myresults into results module and example driver script.

* Untangled methods even more. The goal is for the user to NEVER have to modify the modules.

Properties are now calculated using user specified functions.  Whilst these reside in `model.py`,
the user could specify new property functions and add them to their own module.

I've also taken out the default values that are initialized for the binning and create a smaller function
within examples that specify the binned properties. This hence makes it easier for a user to add new bins.

* The Results class is redundant.

I've moved all the stuff done by the Results class into the example file explicitly.
Have checked that the SMF plots correctly and it does. Now need to update all the other
plots to accept the list of models directly.

* Added docstrings and made the example functions a bit more modular.

* Reworked creation of bins and properties.

The bins and properties are now created in the body of the program explicitly. Makes it clearer what's happening.
Also, bins are nowed stored in Model.bins["bin_name"] to mirror how the properties are done.

* Updated all plots to be compatible with new non-results Class.

This doesn't include TemporalResults (which I shudder to think about just yet...)

* Updated documentation and removed deprecated Results module.

Curious to see if Codacy pops its head in here!

* Add Codacy badge

* Moved Codacy to the badge to the top

* Updated README to reflect new module name.

* Updated history to new format. Die Results Die!

* Started working on using RTD.

* Used the ASTRO3D conf.py

* Maybe the index is pointing to wrong pages.

* Index structure improved. Set up a few skeleton links using Ellert's format.

* Added badge to README.

* Fixed link on index.

* Changed docstrings to use = instead of -

* I'm waging a War on White Space.

* Trying out fsanitize flags.

* Updated Makefile and yml to try get Travis working with fsanitize.

* Tabs in Makefiles make me angry.

* Floating endif was causing flags not to be used.

* Apparently leak is not needed with address.

* Only using fsanitize when specifically asked.

* fsanitize throwing an error.

* Closing atype value. Trying to fix Codacy issues.

* Fixed README badges.

* Getting docs to auto build Docstrings.

* Have to muck around with packages. Things are odd.

* Forgot setup

* Building and everything on my local machine.

* Changed analysis to sage_analysis

* Docs should be compiling now.

* Added an offset for an example and forgot to remove. Whoopsie.

* Trying to get cross-referencing working.

* Added setters for all properties required by Model.

Also added some nifty docstrings!

* Added user html source files.

* Fixed unuses imports and Travis builds.

* Finalised docstrings for Model.

* Fixed up Codacy issues. Broke up the user Documentation.

* Script now preprends the name of the root directory to the OutputDir parameter variable.

* Reworked the flow (again, sigh).

We now use the SAGE parameter file directly, reducing the need to specify a whole bunch of stuff.

As a result, the data subclasses should no longer be subclasses. Instead, they're just normal classes who's __init__
method requires a Model class to call them.

Finally, the OutputDir path for the SAGE parameter file is required to be an absolute path now as a result.

* Updating docs.

* The SimulationDir and FileWithSnapList now required to be absolute paths.

* The plotting now uses classes rather than a subclass.

Both example and history have been updated and checked to run properly.

* Subclass is now a Data Class.

* Sed command needed an extra argument for MacOS.

* Expanded the generate_function_dict function.

It now allows extra keyword arguments to be passed to the function.

* Updated docstring to reflect the new calculation_dict

* Added extra parameter option to selectively not read the SAGE file.

* Changed model_label to just be label.

* Moved calculations/plotting to dedicated modules.

Moved generate_function dict to utils.py.

* Genesis simulation uses the correct number of files.

* Getting rid of white space.

* Split the example and history scripts into a dedicated directory.

* Updated documentation and added new utility function.

* Pass of history to ensure it works properly.

* Updated module level docs.

* Snippet that includes variable about plotting a random subset/all
galaxies.

* Adding pre-commit hooks.

* pre-commit dependency to requirements

* Pre-commit working but not on OzStar.

There's problems with permissions when trying to install a virtual environment
on OzStar.  I've checked on my laptop and it seems things are fine.

* Checking if wget is available. Using curl as fallback.

* Added wget/curl checking to test sage script.

* Added a verbose mode to the sagediff script

* Only check that the substring occurs in the dirname

* Added the verbose mode and removed a couple of not essential print commands

* Fixing issues reported by codacy

* Removed a lot of assert statements

* Fixed some issues pointed out by codacy

* Illustris (#6)

* Fixed one-line return statements

* Fixed the string comparison

* Changed the genesis-lhalo to illustris-lhalo hdf5 reader. Untested

* Made the hdf5 checking a bit more robust

* Fixed the hdf5 field metadata issue

* Fixed crash where 0 galaxies needed to be written out

* Checking if fsanitize runs on TRAVIS

* Removed mem-check

* Fixed the finalize_galaxy bits

* Added illustris-lhalo reader + fixed hdf5 save routines

* Trying to fix tests

* Fixing compile failure on OSX

* Fixing up the LHT hdf5 reader and the save_hdf5

* Increased the chunk size

* Confirmed that snapshot-redshifts and snapnum are being output correctly

* Attempting to make the progressbar better in MPI

* Fixing CTrees reader. ID generation mechanism now set by tree-reader

* Now automatically detecting the hdf5 datatype (and type-checking for every write)

* Fixing up the error messages and the build failure

* Fixing issues raised by codacy

* Added size checking when creating attributes (frac_volume_processed was wrong)

* Now opening the datasets in hdf5 output during the writes

* Replaced asserts with XASSERT

* Made a few more function parameters as const

* Added the two attributes on the master file as well

* Adding a place-holder to relocate tree-level information

* Added the TreeInfo group

* Changed Ngals_per_Snap to be 64-bit integer

* Added the TreeInfo group (and fixed the testing harness)

* Fixed typo in error-checking

* Slightly tweaked the print statement

* Trying to fix codacy warnings

* Final fix for codacy. Remaining warnings are errorneous

* Removed un-necessary indent

* Heh! really should compile before pushing

* Many fixes as per review

* Fixing the outstanding review issues

+++ src/core_save.h
@@ -11,13 +11,13 @@ extern "C" {
                                            struct save_info *save_info,
                                            const struct params *run_params);
 
-    extern int32_t save_galaxies(const int task_forestnr, const int numgals, struct halo_data *halos, struct forest_info *forest_info,
+    extern int32_t save_galaxies(const int64_t task_forestnr, const int numgals, struct halo_data *halos, struct forest_info *forest_info,
                                  struct halo_aux_data *haloaux, struct GALAXY *halogal, struct save_info *save_info,
                                  const struct params *run_params);
 
     extern int32_t finalize_galaxy_files(const struct forest_info *forest_info, struct save_info *save_info,
                                          const struct params *run_params);
-    
+
 #ifdef __cplusplus
 }
 #endif
